clk: qcom: clock-generic: Don't assume parent rate
The try_get_rate flag, when set, allows the mux code to
optimize selecting a parent by checking if a parent is
already at the required rate. However if that parent has
the CLKFLAG_NO_RATE_CACHE flag set, it is possible that
the rate on the parent is stale, and needs to be set again.
Remove the optimization that does an early-exit if the
parent's rate is the requested rate.

Furthermore, some sources may not be able to support
a rate since their own parents may have stale rates.
Confirm that a source can still support a rate by
invoking clk_round_rate on it. The try_get_rate flag
remains useful in real implementations since it
still selects parent already at a given rate that
don't have the restrictions mentioned above.

Change-Id: I4e0c5db934d39ad046755483acf74865b75ca61f
Signed-off-by: Vikram Mulukutla <markivx@codeaurora.org>