msm: qpnp-power-on: support configuring secondary PON for reset
In certain SOCs, multiple PMICs are present to power up the system.
During the system reset, it may not be enough to configure the main
PMIC's PON block based on the power off type. In some cases, there
is a requirement to configure the secondary PMIC's PON block also.
Add support for that.

While at it, fix the error return path so that the driver can probe
properly even when the optional properties are not defined in the
device tree.

CRs-Fixed: 768664
Change-Id: I9741adb76595a3671f766b32ac2d5b16dbdf79ab
Signed-off-by: Subbaraman Narayanamurthy <subbaram@codeaurora.org>