CVE-2016-3857: CONFIG_OABI_COMPAT must be disabled
An elevation of privilege vulnerability in the kernel could enable a
local malicious application to execute arbitrary code within the
context of the kernel.
This issue is rated as Critical due to the possibility of a local
permanent device compromise, which may require reflashing the
operating system to repair the device.

ANDROID_28522518

There is no validation of the events variable passed to the
sys_oabi_epoll_wait function.
The fix is designed to disable OABI support, which will remove the
vulnerable code.

Issue: CYNGNOS-3257

Change-Id: I1002e9feeaecc276aeda73f86ff089b58e9f626f